+++ lib/sidekiq-status/client_middleware.rb
@@ -36,4 +36,15 @@ module Sidekiq::Status
       return msg['args'].to_a.empty? ? nil : msg['args'].to_json
     end
   end
+
+  # Helper method to easily configure sidekiq-status client middleware
+  # whatever the Sidekiq version is.
+  # @param [Sidekiq] sidekiq_config the Sidekiq config
+  # @param [Hash] client_middleware_options client middleware initialization options
+  # @option client_middleware_options [Fixnum] :expiration ttl for complete jobs
+  def self.configure_client_middleware(sidekiq_config, client_middleware_options = {})
+    sidekiq_config.client_middleware do |chain|
+      chain.add Sidekiq::Status::ClientMiddleware, client_middleware_options
+    end
+  end
 end