+++ lib/sidekiq-status/sidekiq_extensions.rb
@@ -1,0 +1,7 @@
+require 'sidekiq/version'
+
+module Sidekiq
+  def self.major_version
+    VERSION.split('.').first.to_i
+  end
+end